Tidied up conversions from real to integer.
git-svn-id: https://scm.physics.ox.ac.uk/svn/radtrancode/trunk@123 4dc0e162-f54d-4b9e-8a18-81c7a7f2ddce

+++ radtran/spec_data/select.f
@@ -242,7 +242,7 @@ C compilers (notably Prospero Fortran for MS-DOS)
       
       
 C     For each bin ...
-      NBIN = (VMAX-VMIN)/BINSIZ + 1
+      NBIN = INT((VMAX-VMIN)/BINSIZ) + 1
       
 C     First setting LSTLIN to one record before the first in the first bin
 C     (FNDWAV returns first record in DBREC) each bin then starts off at one
@@ -283,7 +283,7 @@ C     each strength decade
                   GOTO 110
                ENDIF
 
-               I = ALOG10(LNSTR)
+               I = INT(ALOG10(LNSTR))
                IF(I.GT.MAXSTR)THEN
                   WRITE(*,113)LNWAVE,LNSTR,LNID,LNISO
  113              FORMAT('WARNING - strength too high for storage',
